+++ examples/simple_variable_robot.py
@@ -1,0 +1,25 @@
+from bluedot import BlueDot
+from gpiozero import Robot
+from signal import pause
+
+bd = BlueDot()
+robot = Robot(left=(10, 9), right=(8, 7))
+
+def move(pos):
+    if pos.top:
+        robot.forward(pos.distance)
+    elif pos.bottom:
+        robot.backward(pos.distance)
+    elif pos.left:
+        robot.left(pos.distance)
+    elif pos.right:
+        robot.right(pos.distance)
+
+def stop():
+    robot.stop()
+
+bd.when_pressed = move
+bd.when_moved = move
+bd.when_released = stop
+
+pause()